server_code/requirements.txt
requests==2.32.5
pillow==12.1.0